refactor: Dropdown 컴포넌트와 이벤트 핸들러 -> components의 폴더로
